show my nweets and update displayName

+++ src/components/Nweet.js
@@ -1,5 +1,4 @@
 import { dbService, storageService } from "fbase";
-import { storage } from "firebase";
 import React, { useState } from "react";
 
 const Nweet = ({ nweet, isOwner }) => {
@@ -35,7 +34,9 @@ const Nweet = ({ nweet, isOwner }) => {
   const onDelete = async () => {
     const ok = window.confirm("Are you sure to delete Nweet?");
     if (ok) {
-      await storageService.refFromURL(nweet.attachmentUrl).delete();
+      if (nweet.attachmentUrl !== null) {
+        await storageService.refFromURL(nweet.attachmentUrl).delete();
+      }
       await dbService.doc(`nweets/${nweet.id}`).delete();
     }
   };
